AshGraphql.Resource.Info (ash_graphql v1.3.3)

Introspection helpers for AshGraphql.Resource

Functions

argument_input_types(resource)

Graphql argument type overrides for the resource

argument_names(resource)

Graphql argument name overrides for the resource

attribute_input_types(resource)

Graphql attribute input type overrides for the resource

attribute_types(resource)

Graphql type overrides for the resource

derive_filter?(resource)

Wether or not to derive a filter input for the resource automatically

derive_sort?(resource)

Wether or not to derive a sort input for the resource automatically

encode_primary_key?(resource)

Wether or not to encode the primary key as a single id field when reading and getting

field_names(resource)

Graphql field name (attribute/relationship/calculation/arguments) overrides for the resource

filterable_field?(resource, field_name)

May the specified field be filtered on?

filterable_fields(resource)

Fields that may be filtered on

generate_object?(resource)

Wether or not an object should be generated, or if one with the type name for this resource should be used.

hide_fields(resource)

Fields to hide from the graphql domain

keyset_field(resource)

The field name to place the keyset of a result in

managed_relationship(resource, action, argument)

The managed_relationshi configuration for a given action/argument

managed_relationships(resource)

The managed_relationship configurations

managed_relationships_auto?(resource)

mutations(resource, domain_or_domains \\ [])

The mutations exposed for the resource

nullable_fields(resource)

Graphql nullability overrides for the resource

paginate_relationship_with(resource)

Pagination configuration for list relationships

primary_key_delimiter(resource)

The delimiter for a resource with a composite primary key

queries(resource, domain_or_domains \\ [])

The queries exposed for the resource

relationships(resource)

Which relationships should be included in the generated type

show_field?(resource, field)

Wether or not a given field will be shown

show_fields(resource)

Fields to show in the graphql domain

type(resource)

The graphql type of the resource